pyproject.toml
[project]
name = "typecorrect"
version = "0.1.0"
description = "Enforces the correct use of the left and right Shift keys based on keyboard layout."
readme = "readme.md"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=11.0.0",
    "pynput>=1.8.0",
    "pystray>=0.19.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.0.0",
]
